mito-ai: create_validate_streamlit_app handler
